Update logical functions to handle change
The change of placing `struct list_head` only in free pages
means that calculation needs to refer calculate virtual addresses
(I think) rather than offsets in `struct hyp_page` in the
vmemmap. `init_vmemmap` has its own comment - it is simplified
because (1) the output arguments requires are not available and
(2) the condition being commented out is being checked in an invariant
at the call site.