Make progress on __hyp_attach_page
Basically, the lack of being able to name and pass around ghost
parameters is the root of the issue.  If this were expressible, this
issue wouldn't arise in the first place.

1) To express something like cellPointer(hyp_vmemmap, 4, start_i, end_i,
p), I need let O = Owned(pool) for pool->range_{start_end}

1a) However, this gives me ownership of pool->free_area[MAX_ORDER] as
well (which is acceptable)

2) But this means I can't straightforwardly pass struct list_head* head
as a parameter, or have let O2 = Owned(head) as a pre-condition (which
is necessary for the spec of page_add_to_list).

2a) If I do, then the resource is missing - and the lack of info on
where a resource is consumed means that figuring this out was way more
time consuming than it needs to be

2b) I don't think I can split up pool to pass `.range_start, .range_end,
.max_order` because of `hyp_pool_wf(pool, *pool, ..)`