Bitvectors: fix __find_buddy_nocheck
The previous elaboration of `<<`/shift-left in CN used the `power_uf`
SMT term, however, with bitvectors it uses the newly added `shift_left`.
This led to a mismatch between the old specification (which still used
`power_uf` and the new elaborated symbolic expression which used
`shift_left` for the statement `addr ^= PAGE_SIZE << order;`.

The fix is to stop using `power_uf` and use `shift_left` instead.